+++ src/platform/admin/licensing/admin.licensing.resolver.mutations.ts
@@ -5,11 +5,11 @@ import { GraphqlGuard } from '@core/authorization';
 import { AgentInfo } from '@core/authentication.agent.info/agent.info';
 import { AuthorizationService } from '@core/authorization/authorization.service';
 import { AuthorizationPrivilege } from '@common/enums/authorization.privilege';
-import { AssignLicensePlanToSpace } from './dto/admin.licensing.dto.assign.license.plan.to.account';
+import { AssignLicensePlanToSpace } from './dto/admin.licensing.dto.assign.license.plan.to.space';
 import { LicensingService } from '@platform/licensing/licensing.service';
 import { ILicensing } from '@platform/licensing/licensing.interface';
 import { AdminLicensingService } from './admin.licensing.service';
-import { RevokeLicensePlanFromSpace } from './dto/admin.licensing.dto.revoke.license.plan.from.account';
+import { RevokeLicensePlanFromSpace } from './dto/admin.licensing.dto.revoke.license.plan.from.space';
 import { AuthorizationPolicyService } from '@domain/common/authorization-policy/authorization.policy.service';
 import { ISpace } from '@domain/space/space/space.interface';
 import { SpaceAuthorizationService } from '@domain/space/space/space.service.authorization';
