+++ src/platform/admin/licensing/dto/admin.licensing.dto.assign.license.plan.to.space.ts
