+++ src/platform/admin/licensing/dto/admin.licensing.dto.revoke.license.plan.from.space.ts
